progress
"resending email verification > email is resent correctly" is being weird.
If I run all the tests, it fails. If I ran all the tests except it, they
all pass. If I run it on its own, it passes. So I think something in the test
is somehow conflicting with something in the other tests, e.g. UserModels,
but I have no idea how since everything's random and the database gets
cleared between test runs and all the tests run sequentially.