Fix npm build to include email templates properly
Due to the slashes placed in the build script around `templates`,
the build script would copy the entire directory directly into the
existing `build/templates` folder from the Docker cache, which is not
what we want for reproducible builds. This caused non-cached builds
to not properly copy over email templates required for the program.

Fix by copying the entire directory into the build directory, and NOT
into the `build/templates` directory.

`cp` is weird with slashes, which is what made the difference here.
Basically, adding the slash to the end of `build/templates/` caused
the entire folder itself to be copied inside, causing unnecessary
nesting (final result was `templates -> build/templates/templates`).